Changed Rigidbody to CharacterController
I'd like to get more familiar with things like characterController; I
ignore prebuilt solutions to a fault, so I want to do something to speed
up dev a little...

Also added a couple things to gitignore that VS was generating that I
didn't really want to add to this.  I'll undo them if that's a problem.